feat: v2 refactor
