+++ lua/git-worktree/init.lua
@@ -10,140 +10,6 @@ local git_worktree_root = nil
 local current_worktree_path = nil
 local on_change_callbacks = {}
 
-M.setup_git_info = function()
-    local cwd = vim.loop.cwd()
-
-    local is_in_worktree = false
-
-    local inside_worktree_job = Job:new({
-        'git', 'rev-parse', '--is-inside-work-tree',
-        cwd = cwd,
-    })
-
-    local process_inside_worktree = function(stdout)
-        if stdout == "true" then
-            is_in_worktree = true
-        end
-    end
-
-    local find_git_dir_job = Job:new({
-        'git', 'rev-parse', '--absolute-git-dir',
-        cwd = cwd,
-    })
-
-    local process_find_git_dir = function(stdout)
-        if is_in_worktree then
-            -- if in worktree git dir returns absolute path
-
-            -- try to find the dot git folder (non-bare repo)
-            local git_dir = Path:new(stdout)
-            local has_dot_git = false
-            for _, dir in ipairs(git_dir:_split()) do
-                if dir == ".git" then
-                    has_dot_git = true
-                    break
-                end
-            end
-
-            if has_dot_git then
-                if stdout == ".git" then
-                    git_worktree_root = cwd
-                else
-                    local start = stdout:find("%.git")
-                    git_worktree_root = stdout:sub(1,start - 2)
-                end
-            else
-                local start = stdout:find("/worktrees/")
-                git_worktree_root = stdout:sub(0, start - 1)
-            end
-        elseif stdout == "." then
-            -- we are in the root git dir
-            git_worktree_root = cwd
-        else
-            -- if not in worktree git dir should be absolute
-            git_worktree_root = stdout
-        end
-        status:log():debug("git directory is: " .. git_worktree_root)
-    end
-
-    local find_toplevel_job = Job:new({
-        'git', 'rev-parse', '--show-toplevel',
-        cwd = cwd,
-    })
-
-    local find_toplevel_bare_job = Job:new({
-        'git', 'rev-parse', '--is-bare-repository',
-        cwd = cwd,
-    })
-
-    local process_find_toplevel = function(stdout)
-        current_worktree_path = stdout
-        status:log().debug("git toplevel is: " .. current_worktree_path)
-    end
-
-    local process_find_toplevel_bare = function(stdout)
-        if stdout == "true" then
-            current_worktree_path = cwd
-        end
-    end
-
-    local stdout, code = inside_worktree_job:sync()
-    if code ~= 0 then
-        status:log().error("Error in determining if we are in a worktree")
-        git_worktree_root = nil
-        current_worktree_path = nil
-        return
-    end
-    stdout = table.concat(stdout, "")
-    process_inside_worktree(stdout)
-
-    stdout, code = find_git_dir_job:sync()
-    if code ~= 0 then
-        status:log().error("Error in determining the git root dir")
-        git_worktree_root = nil
-        return
-    end
-    stdout = table.concat(stdout, "")
-    process_find_git_dir(stdout)
-
-    stdout, code = find_toplevel_job:sync()
-    if code == 0 then
-        stdout = table.concat(stdout, "")
-        process_find_toplevel(stdout)
-    else
-        stdout, code = find_toplevel_bare_job:sync()
-        if code == 0 then
-            stdout = table.concat(stdout, "")
-            process_find_toplevel_bare(stdout)
-        else
-            status:log().error("Error in determining the git toplevel")
-            current_worktree_path = nil
-            return
-        end
-    end
-
-end
-
-local function on_tree_change_handler(op, metadata)
-    if M._config.update_on_change then
-        if op == Enum.Operations.Switch then
-            local changed = M.update_current_buffer(metadata["prev_path"])
-            if not changed then
-                status:log().debug("Could not change to the file in the new worktree, running the `update_on_change_command`")
-                vim.cmd(M._config.update_on_change_command)
-            end
-        end
-    end
-end
-
-local function emit_on_change(op, metadata)
-    -- TODO: We don't have a way to async update what is running
-    status:next_status(string.format("Running post %s callbacks", op))
-    on_tree_change_handler(op, metadata)
-    for idx = 1, #on_change_callbacks do
-        on_change_callbacks[idx](op, metadata)
-    end
-end
 
 local function change_dirs(path)
     local worktree_path = M.get_worktree_path(path)
@@ -157,7 +23,7 @@ local function change_dirs(path)
         vim.cmd(cmd)
         current_worktree_path = worktree_path
     else
-        status:error('Could not chang to directory: ' ..worktree_path)
+        status:error('Could not chang to directory: ' .. worktree_path)
     end
 
     if M._config.clearjumps_on_change then
@@ -169,9 +35,8 @@ local function change_dirs(path)
 end
 
 local function create_worktree_job(path, branch, found_branch)
-
     local worktree_add_cmd = 'git'
-    local worktree_add_args = {'worktree', 'add'}
+    local worktree_add_args = { 'worktree', 'add' }
 
     if not found_branch then
         table.insert(worktree_add_args, '-b')
@@ -199,8 +64,10 @@ local function has_worktree(path, cb)
     local plenary_path = Path:new(path)
 
     local job = Job:new({
-        'git', 'worktree', 'list', on_stdout = function(_, data)
-
+        'git',
+        'worktree',
+        'list',
+        on_stdout = function(_, data)
             local list_data = {}
             for section in data:gmatch("%S+") do
                 table.insert(list_data, section)
@@ -253,44 +120,7 @@ local function failure(from, cmd, path, soft_error)
     end
 end
 
-local function has_origin()
-    local found = false
-    local job = Job:new({
-        'git', 'remote', 'show',
-        on_stdout = function(_, data)
-            data = vim.trim(data)
-            found = found or data == 'origin'
-        end,
-        cwd = git_worktree_root,
-    })
 
-    -- TODO: I really don't want status's spread everywhere... seems bad
-    job:after(function()
-        status:status("found origin: " .. tostring(found))
-    end):sync()
-
-    return found
-end
-
-local function has_branch(branch, cb)
-    local found = false
-    local job = Job:new({
-        'git', 'branch', on_stdout = function(_, data)
-            -- remove  markere on current branch
-            data = data:gsub("*","")
-            data = vim.trim(data)
-            found = found or data == branch
-        end,
-        cwd = git_worktree_root,
-    })
-
-    -- TODO: I really don't want status's spread everywhere... seems bad
-    status:next_status(string.format("Checking for branch %s", branch))
-    job:after(function()
-        status:status("found branch: " .. tostring(found))
-        cb(found)
-    end):start()
-end
 
 local function create_worktree(path, branch, upstream, found_branch)
     local create = create_worktree_job(path, branch, found_branch)
@@ -302,17 +132,19 @@ local function create_worktree(path, branch, upstream, found_branch)
         worktree_path = Path:new(git_worktree_root, path):absolute()
     end
 
-    local fetch = Job:new({
-        'git', 'fetch', '--all',
+    local fetch           = Job:new({
+        'git',
+        'fetch',
+        '--all',
         cwd = worktree_path,
         on_start = function()
             status:next_status("git fetch --all (This may take a moment)")
         end
     })
 
-    local set_branch_cmd = 'git'
-    local set_branch_args= {'branch', string.format('--set-upstream-to=%s/%s', upstream, branch)}
-    local set_branch = Job:new({
+    local set_branch_cmd  = 'git'
+    local set_branch_args = { 'branch', string.format('--set-upstream-to=%s/%s', upstream, branch) }
+    local set_branch      = Job:new({
         command = set_branch_cmd,
         args = set_branch_args,
         cwd = worktree_path,
@@ -323,9 +155,9 @@ local function create_worktree(path, branch, upstream, found_branch)
 
     -- TODO: How to configure origin???  Should upstream ever be the push
     -- destination?
-    local set_push_cmd = 'git'
-    local set_push_args = {'push', "--set-upstream", upstream,  branch, path}
-    local set_push  = Job:new({
+    local set_push_cmd    = 'git'
+    local set_push_args   = { 'push', "--set-upstream", upstream, branch, path }
+    local set_push        = Job:new({
         command = set_push_cmd,
         args = set_push_args,
         cwd = worktree_path,
@@ -334,8 +166,9 @@ local function create_worktree(path, branch, upstream, found_branch)
         end
     })
 
-    local rebase = Job:new({
-        'git', 'rebase',
+    local rebase          = Job:new({
+        'git',
+        'rebase',
         cwd = worktree_path,
         on_start = function()
             status:next_status("git rebase")
@@ -362,20 +195,19 @@ local function create_worktree(path, branch, upstream, found_branch)
         set_branch:after_failure(failure("create_worktree", set_branch.args, worktree_path, true))
 
         rebase:after(function()
-
             if rebase.code ~= 0 then
                 status:status("Rebase failed, but that's ok.")
             end
 
             vim.schedule(function()
-                emit_on_change(Enum.Operations.Create, {path = path, branch = branch, upstream = upstream})
+                emit_on_change(Enum.Operations.Create, { path = path, branch = branch, upstream = upstream })
                 M.switch_worktree(path)
             end)
         end)
     else
         create:after(function()
             vim.schedule(function()
-                emit_on_change(Enum.Operations.Create, {path = path, branch = branch, upstream = upstream})
+                emit_on_change(Enum.Operations.Create, { path = path, branch = branch, upstream = upstream })
                 M.switch_worktree(path)
             end)
         end)
@@ -404,14 +236,12 @@ M.create_worktree = function(path, branch, upstream)
             create_worktree(path, branch, upstream, found_branch)
         end)
     end)
-
 end
 
 M.switch_worktree = function(path)
     status:reset(2)
     M.setup_git_info()
     has_worktree(path, function(found)
-
         if not found then
             status:error("worktree does not exists, please create it first " .. path)
         end
@@ -420,7 +250,6 @@ M.switch_worktree = function(path)
             local prev_path = change_dirs(path)
             emit_on_change(Enum.Operations.Switch, { path = path, prev_path = prev_path })
         end)
-
     end)
 end
 
@@ -485,7 +314,7 @@ M.update_current_buffer = function(prev_path)
     end
 
     local name = Path:new(current_buf_name):absolute()
-    local start, fin = string.find(name, cwd..Path.path.sep, 1, true)
+    local start, fin = string.find(name, cwd .. Path.path.sep, 1, true)
     if start ~= nil then
         return true
     end
@@ -497,7 +326,7 @@ M.update_current_buffer = function(prev_path)
 
     local local_name = name:sub(fin + 2)
 
-    local final_path = Path:new({cwd, local_name}):absolute()
+    local final_path = Path:new({ cwd, local_name }):absolute()
 
     if not Path:new(final_path):exists() then
         return false
